+++ src/main/java/shop/wannab/book_service/book/service/AdminBookService.java
@@ -1,5 +1,6 @@
 package shop.wannab.book_service.book.service;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -7,7 +8,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shop.wannab.book_service.author.entity.Author;
 import shop.wannab.book_service.author.repository.AuthorRepository;
-import shop.wannab.book_service.book.controller.request.AladinBookCreateRequest;
 import shop.wannab.book_service.book.controller.request.BookCreateRequest;
 import shop.wannab.book_service.book.controller.request.BookUpdateRequest;
 import shop.wannab.book_service.book.controller.response.BookListResponse;
@@ -20,8 +20,6 @@ import shop.wannab.book_service.book.exception.BookErrorCode;
 import shop.wannab.book_service.book.repository.BookRepository;
 import shop.wannab.book_service.publisher.entity.Publisher;
 import shop.wannab.book_service.publisher.repository.PublisherRepository;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -146,47 +144,4 @@ public class AdminBookService {
         bookRepository.deleteBookStock(bookId);
     }
 
-    public void aladinCreateBook(AladinBookCreateRequest request) {
-
-        if (bookRepository.existsByIsbn(request.isbn())) {
-            throw new BookApiException(BookErrorCode.DUPLICATE_BOOK);
-        }
-
-        Book book = request.toEntity();
-        List<BookAuthor> bookAuthors = request.authors().stream()
-                .map(authorName ->{
-                    Author author = authorRepository.findAuthorsByAuthorName(authorName)
-                            .orElseGet(()->authorRepository.save(
-                                    Author.builder().authorName(authorName).build()));
-                    return BookAuthor.builder()
-                            .book(book)
-                            .author(author)
-                            .build();
-                }).toList();
-
-        List<BookPublisher> bookPublishers = request.publishers().stream()
-                .map(publisherName -> {
-                    Publisher publisher = publisherRepository.findPublisherByPublisherName(publisherName)
-                            .orElseGet(() -> publisherRepository.save(
-                                    Publisher.builder().publisherName(publisherName).build()));
-                    return BookPublisher.builder()
-                            .book(book)
-                            .publisher(publisher)
-                            .build();
-                }).toList();
-
-        List<BookImage> bookImages = request.thumbnail().stream()
-                .map(imageUrl -> BookImage.builder()
-                        .book(book)
-                        .imageUrl(imageUrl)
-                        .build())
-                .toList();
-
-        book.getBookImages().addAll(bookImages);
-        book.getBookAuthors().addAll(bookAuthors);
-        book.getBookPublishers().addAll(bookPublishers);
-
-        bookRepository.save(book);
-        bookRepository.saveOrUpdateBookStock(book.getBookId(),book.getStock());
-    }
 }
